restructure and adding of Round

+++ Services/QuestionService/QuestionService.cs
@@ -1,8 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using QuizFlow.Data;
 using QuizFlow.Dto.Question;
@@ -13,68 +15,59 @@ namespace QuizFlow.Services.QuestionService {
 
     private readonly IMapper _mapper;
     private readonly DataContext _dbContext;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
-    public QuestionService(IMapper mapper, DataContext context) {
-      _dbContext = context;
+    public QuestionService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor) {
       _mapper = mapper;
+      _dbContext = context;
+      _httpContextAccessor = httpContextAccessor;
     }
 
     public async Task<ServiceResponse<List<QuestionDtoGet>>> getAllQuestions() {
-      // create res
       ServiceResponse<List<QuestionDtoGet>> serviceResponse = new ServiceResponse<List<QuestionDtoGet>>();
-      // get data from DB
       List<Question> dbQuestions = await _dbContext.Questions.ToListAsync();
-      // add data to res
       serviceResponse.data = dbQuestions.Select(q => _mapper.Map<QuestionDtoGet>(q)).ToList();
-      // return res
       return serviceResponse;
     }
 
     public async Task<ServiceResponse<QuestionDtoGet>> getQuestionById(int id) {
-      // create res
       ServiceResponse<QuestionDtoGet> serviceResponse = new ServiceResponse<QuestionDtoGet>();
-      // get data from DB
-      Question question = await _dbContext.Questions.FirstOrDefaultAsync(q => q.id == id);
-      // add data to res
+      Question question = await _dbContext.Questions.FirstOrDefaultAsync(q => q.id == id && q.user.id == getUserId());
       serviceResponse.data = _mapper.Map<QuestionDtoGet>(question);
-      // return res
       return serviceResponse;
     }
 
-    public async Task<ServiceResponse<List<QuestionDtoGet>>> getUserQuestions(int id) {
-      // create res
+    public async Task<ServiceResponse<List<QuestionDtoGet>>> getUserQuestions() {
       ServiceResponse<List<QuestionDtoGet>> serviceResponse = new ServiceResponse<List<QuestionDtoGet>>();
-      // get users data from DB
-      List<Question> dbQuestions = await _dbContext.Questions.Where(q => q.user.id == id).ToListAsync();
-      // add data to res
+      List<Question> dbQuestions = await _dbContext.Questions.Where(q => q.user.id == getUserId()).ToListAsync();
       serviceResponse.data = dbQuestions.Select(q => _mapper.Map<QuestionDtoGet>(q)).ToList();
-      // return res
       return serviceResponse;
     }
 
     public async Task<ServiceResponse<QuestionDtoGet>> addQuestion(QuestionDtoAdd newQuestion) {
-      // create res
       ServiceResponse<QuestionDtoGet> serviceResponse = new ServiceResponse<QuestionDtoGet>();
-      // add data to DB
       Question question = _mapper.Map<Question>(newQuestion);
+      question.user = await _dbContext.Users.FirstOrDefaultAsync(u => u.id == getUserId());
       await _dbContext.Questions.AddAsync(question);
       await _dbContext.SaveChangesAsync();
-      // add data to res
       serviceResponse.data = _mapper.Map<QuestionDtoGet>(question);
-      // return res
       return serviceResponse;
     }
 
     public async Task<ServiceResponse<QuestionDtoGet>> editQuestion(QuestionDtoEdit editedQuestion) {
       ServiceResponse<QuestionDtoGet> serviceResponse = new ServiceResponse<QuestionDtoGet>();
       try {
-        Question question = await _dbContext.Questions.FirstOrDefaultAsync(q => q.id == editedQuestion.id);
-        question.question = editedQuestion.question;
-        question.answer = editedQuestion.answer;
-        // must set all values manually or will change to default values
-        _dbContext.Questions.Update(question);
-        await _dbContext.SaveChangesAsync();
-        serviceResponse.data = _mapper.Map<QuestionDtoGet>(question);
+        Question question = await _dbContext.Questions.Include(q => q.user).FirstOrDefaultAsync(q => q.id == editedQuestion.id);
+        if (question.user.id == getUserId()) {
+          question.question = editedQuestion.question;
+          question.answer = editedQuestion.answer;
+          _dbContext.Questions.Update(question);
+          await _dbContext.SaveChangesAsync();
+          serviceResponse.data = _mapper.Map<QuestionDtoGet>(question);
+        } else {
+          serviceResponse.success = false;
+          serviceResponse.message = "User question not found";
+        }
       } catch (Exception e) {
         serviceResponse.success = false;
         serviceResponse.message = e.Message;
@@ -85,15 +78,24 @@ namespace QuizFlow.Services.QuestionService {
     public async Task<ServiceResponse<string>> deleteQuestion(int id) {
       ServiceResponse<string> serviceResponse = new ServiceResponse<string>();
       try {
-        Question question = await _dbContext.Questions.FirstAsync(q => q.id == id);
-        _dbContext.Questions.Remove(question);
-        await _dbContext.SaveChangesAsync();
-        serviceResponse.data = "success";
+        Question question = await _dbContext.Questions.FirstOrDefaultAsync(q => q.id == id && q.user.id == getUserId());
+        if (question != null) {
+          _dbContext.Questions.Remove(question);
+          await _dbContext.SaveChangesAsync();
+          serviceResponse.data = "success";
+        } else {
+          serviceResponse.success = false;
+          serviceResponse.message = "User question not found";
+        }
       } catch (Exception e) {
         serviceResponse.success = false;
         serviceResponse.message = e.Message;
       }
       return serviceResponse;
+    }
+
+    private int getUserId() {
+      return int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
     }
 
   }
